+++ packages/providers-evm/src/base-evm.ts
@@ -34,15 +34,13 @@ export class BaseEVMProvider extends ProviderEventBus implements ProviderProxy {
   constructor(provider?: RawProvider | providers.Web3Provider) {
     if (!provider) throw new errors.ProviderInjectedInstanceNotFoundError()
     super()
-    this.#provider =
+    const rawProvider = wrapExternalEthProvider(
       provider instanceof providers.Web3Provider
-        ? provider
-        : new ethers.providers.Web3Provider(
-            wrapExternalEthProvider(
-              provider as ethers.providers.ExternalProvider,
-            ),
-            'any',
-          )
+        ? provider.provider
+        : (provider as providers.ExternalProvider),
+    )
+
+    this.#provider = new ethers.providers.Web3Provider(rawProvider, 'any')
   }
 
   get chainType(): ChainTypes {